Adjust the API available from NPM.
Previously, you could only get access to an instance of the client when using NPM.  This was a little less than wonderful given for things like testing, where you might want to be able to fire up a coordinator locally.  This change adjusts the API to make the server programmatically instantiatable as well.